requirements.txt
altair
numpy
pandas
pydeck
streamlit
openpyxl
plotly.express